Fix warnings from -Wstrict-overflow and -Wconversion.
git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@3712 95820547-d848-0410-985e-9ae8fe0fa350

+++ libelf/elf_end.c
@@ -81,7 +81,7 @@ elf_end(Elf *e)
 				free(e->e_rawfile);
 #if	ELFTC_HAVE_MMAP
 			else if (e->e_flags & LIBELF_F_RAWFILE_MMAP)
-				(void) munmap(e->e_rawfile, e->e_rawsize);
+				(void) munmap(e->e_rawfile, (size_t) e->e_rawsize);
 #endif
 		}
